Rename `test_config.rs` to `build_config.rs`
To match the rename of `TestConfig` to `BuildConfig`.

Follow-up to #470.

GUS-W-11467987.
